# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Custom Nails API
  description: |
    REST API for custom nail design orders, allowing customers to submit designs and technicians to review and fulfill orders.

    ## Authentication
    All protected endpoints require a valid JWT token from Auth0 in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## User Roles
    - **customer**: Can create orders and view their own orders
    - **technician**: Can view unassigned orders, assign orders to themselves, review orders, and update order status
  version: 1.0.0
  contact:
    name: API Support
    email: support@kendallsnails.com
  license:
    name: MIT

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.kendallsnails.com/api/v1
    description: Production server

tags:
  - name: Health
    description: Health check and system status endpoints
  - name: Users
    description: User management endpoints
  - name: Orders
    description: Order management endpoints

paths:
  /health:
    get:
      summary: Health check
      description: Returns the health status of the API
      operationId: getHealth
      tags:
        - Health
      responses:
        '200':
          description: API is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Custom Nails API is running"

  /database/status:
    get:
      summary: Database status
      description: Returns the database connection status and list of tables
      operationId: getDatabaseStatus
      tags:
        - Health
      responses:
        '200':
          description: Database is connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Database connected"
                  tables:
                    type: array
                    items:
                      type: string
                    example: ["users", "orders"]

  /users:
    post:
      summary: Create user
      description: Create a new user profile (typically called after Auth0 registration)
      operationId: createUser
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                role:
                  type: string
                  enum: [customer, technician]
                  default: customer
                  example: "customer"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/me:
    get:
      summary: Get current user profile
      description: Retrieve the profile of the authenticated user
      operationId: getMyProfile
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/UserNotFound'

    put:
      summary: Update current user profile
      description: Update the profile of the authenticated user
      operationId: updateMyProfile
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "John Doe"
                email:
                  type: string
                  format: email
                  example: "john@example.com"
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/UserNotFound'

  /orders:
    post:
      summary: Create order
      description: Create a new custom nail design order (customers only)
      operationId: createOrder
      tags:
        - Orders
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
                - quantity
              properties:
                description:
                  type: string
                  example: "Pink nails with glitter"
                quantity:
                  type: integer
                  minimum: 1
                  example: 2
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only customers can create orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List orders
      description: |
        List orders with role-based filtering:
        - Customers see only their own orders
        - Technicians see orders assigned to them + unassigned orders
      operationId: listOrders
      tags:
        - Orders
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of orders per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /orders/{id}:
    get:
      summary: Get order
      description: |
        Retrieve a single order by ID with authorization:
        - Customers can only access their own orders
        - Technicians can access orders assigned to them or unassigned orders
      operationId: getOrder
      tags:
        - Orders
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      responses:
        '200':
          description: Order retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: You do not have permission to access this order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/OrderNotFound'

  /orders/{id}/assign:
    put:
      summary: Assign order to technician
      description: Assign an order to the current technician (technicians only)
      operationId: assignOrder
      tags:
        - Orders
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      responses:
        '200':
          description: Order assigned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only technicians can assign orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '422':
          description: Order is already assigned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: "ALREADY_ASSIGNED"
                      message:
                        type: string
                        example: "Order is already assigned to another technician"

  /orders/{id}/review:
    put:
      summary: Review order
      description: Accept or reject an order (technicians only)
      operationId: reviewOrder
      tags:
        - Orders
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required:
                    - action
                    - price
                  properties:
                    action:
                      type: string
                      enum: [accept]
                      example: "accept"
                    price:
                      type: number
                      format: float
                      minimum: 0
                      exclusiveMinimum: true
                      example: 45.00
                - type: object
                  required:
                    - action
                    - feedback
                  properties:
                    action:
                      type: string
                      enum: [reject]
                      example: "reject"
                    feedback:
                      type: string
                      example: "Design is too complex for current materials"
      responses:
        '200':
          description: Order reviewed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only technicians can review orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '422':
          description: Order has already been reviewed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: "INVALID_STATE"
                      message:
                        type: string
                        example: "Order has already been reviewed"

  /orders/{id}/status:
    put:
      summary: Update order status
      description: |
        Update the status of an order through the production workflow (technicians only).

        Valid status transitions:
        - `accepted` → `in_production`
        - `in_production` → `shipped`
        - `shipped` → `delivered`

        The order must be assigned to the requesting technician.
      operationId: updateOrderStatus
      tags:
        - Orders
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [in_production, shipped, delivered]
                  description: The new status for the order
                  example: "in_production"
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
              example:
                success: true
                data:
                  id: 1
                  description: "Pink nails with glitter"
                  quantity: 2
                  status: "in_production"
                  price: 45.00
                  feedback: null
                  customer_id: 1
                  technician_id: 2
                  created_at: "2025-01-15T10:30:00Z"
                  updated_at: "2025-01-15T11:00:00Z"
                  customer:
                    id: 1
                    name: "Jane Customer"
                    email: "jane@example.com"
                    role: "customer"
                  technician:
                    id: 2
                    name: "Tech Smith"
                    email: "tech@example.com"
                    role: "technician"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only technicians can update order status, or order is not assigned to requesting technician
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum: [FORBIDDEN]
                        example: "FORBIDDEN"
                      message:
                        type: string
                        example: "Only technicians can update order status"
              examples:
                notTechnician:
                  value:
                    success: false
                    error:
                      code: "FORBIDDEN"
                      message: "Only technicians can update order status"
                notAssigned:
                  value:
                    success: false
                    error:
                      code: "FORBIDDEN"
                      message: "You can only update status of orders assigned to you"
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '422':
          description: Invalid status transition
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum: [INVALID_STATE, INVALID_TRANSITION]
                        example: "INVALID_TRANSITION"
                      message:
                        type: string
                        example: "Invalid status transition"
                      details:
                        type: object
                        properties:
                          current_status:
                            type: string
                            example: "accepted"
                          requested_status:
                            type: string
                            example: "shipped"
                          allowed_statuses:
                            type: array
                            items:
                              type: string
                            example: ["in_production"]
              examples:
                invalidTransition:
                  summary: Trying to skip a status (e.g., accepted → shipped)
                  value:
                    success: false
                    error:
                      code: "INVALID_TRANSITION"
                      message: "Invalid status transition"
                      details:
                        current_status: "accepted"
                        requested_status: "shipped"
                        allowed_statuses: ["in_production"]
                invalidState:
                  summary: Trying to update from an invalid state (e.g., submitted)
                  value:
                    success: false
                    error:
                      code: "INVALID_STATE"
                      message: "Cannot update status from current order state"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 JWT token

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        auth0_id:
          type: string
          example: "auth0|123456789"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [customer, technician]
          example: "customer"
        created_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"

    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        description:
          type: string
          example: "Pink nails with glitter"
        quantity:
          type: integer
          example: 2
        status:
          type: string
          enum: [submitted, accepted, rejected, in_production, shipped, delivered]
          example: "submitted"
        price:
          type: number
          format: float
          nullable: true
          example: 45.00
        feedback:
          type: string
          nullable: true
          example: "Design is too complex for current materials"
        customer_id:
          type: integer
          example: 1
        customer:
          $ref: '#/components/schemas/User'
        technician_id:
          type: integer
          nullable: true
          example: 2
        technician:
          allOf:
            - $ref: '#/components/schemas/User'
          nullable: true
        created_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"

    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 25
        totalPages:
          type: integer
          example: 3

    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "ERROR_CODE"
            message:
              type: string
              example: "Error message"
            details:
              type: object
              nullable: true

  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT token
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "UNAUTHORIZED"
                  message:
                    type: string
                    example: "Could not extract user information"

    ValidationError:
      description: Validation error - Invalid request data
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "VALIDATION_ERROR"
                  message:
                    type: string
                    example: "Invalid request data"
                  details:
                    type: string
                    example: "Field validation error details"

    UserNotFound:
      description: User not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "USER_NOT_FOUND"
                  message:
                    type: string
                    example: "User profile not found. Please create a profile first."

    OrderNotFound:
      description: Order not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  code:
                    type: string
                    example: "ORDER_NOT_FOUND"
                  message:
                    type: string
                    example: "Order not found"